+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 0.1.3
 - Deprecate Python 3.7, add Python 3.12
+- Fix executor's shutdown
 
 # 0.1.2
 - Just more user-friendly readme & library description: links, classifiers, keywords